fix(bsp-esp32): repair cross-build after bt-module vendoring
The esp-idf-svc vendoring let the build reach bsp-esp32 for the first time,
exposing compile errors in the never-built hardware modules:
- ble: init controller via esp_bt_controller_get_initial_config (the
  BT_CONTROLLER_INIT_CONFIG_DEFAULT macro is not bound), pass *mut pointers
  to config_adv_data_raw/start_advertising, and move the BLE-5 extended-adv
  body under cfg(s3/c6) so classic esp32 no longer compiles the missing
  PHY/EXT_ADV_PROP enum symbols.
- usb: gate the module in lib.rs to s3/c6 (inner cfg did not exclude it on
  esp32) and cast length to u32 for usb_serial_jtag_read_bytes.
- ota: fix esp_ota_begin image_size to usize, drop unneeded mut.
- nvs: fix double mutable borrow in get_str with split_at_mut.
- led: wrap mutable-static raw pointers in the unsafe block.
- web: pass Arc<SharedState> into handlers instead of a raw *const (Send).
- wifi: drop unused imports.

+++ OmniRID/hardware/bsp-esp32/src/usb.rs
@@ -5,9 +5,7 @@
 //! for incoming MAVLink data from the flight controller.
 //!
 //! The USB Serial/JTAG peripheral only exists on ESP32-S3/C6 (not classic
-//! ESP32), so the whole module is only compiled for those targets.
-
-#![cfg(any(feature = "esp32s3", feature = "esp32c6"))]
+//! ESP32), so `lib.rs` only declares this module for those targets.
 
 use esp_idf_svc as _;
 use esp_idf_svc::sys::{self as sys};
@@ -43,7 +41,7 @@ impl UartRead for UsbCdcRx {
         unsafe {
             let n = sys::usb_serial_jtag_read_bytes(
                 buf.as_mut_ptr() as *mut _,
-                buf.len(),
+                buf.len() as u32,
                 0, // non-blocking
             );
             if n > 0 { n as usize } else { 0 }
